requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
openai==1.58.1
python-telegram-bot==21.9
httpx==0.28.1
python-dotenv==1.0.1
bcrypt==4.2.1
itsdangerous==2.2.0
python-multipart==0.0.20
